backend/requirements.txt
flask
flask-cors
requests
azure-ai-textanalytics==5.2.0